add deleteObject in jpa removed unused classes

+++ jdy.model.jpa/src/main/java/de/jdynameta/jdy/model/jpa/JpaWriter.java
@@ -3,13 +3,18 @@ package de.jdynameta.jdy.model.jpa;
 import de.jdynameta.base.metainfo.AttributeHandler;
 import de.jdynameta.base.metainfo.ObjectReferenceAttributeInfo;
 import de.jdynameta.base.metainfo.PrimitiveAttributeInfo;
+import de.jdynameta.base.metainfo.filter.ObjectFilterExpression;
+import de.jdynameta.base.metainfo.filter.defaultimpl.*;
 import de.jdynameta.base.value.JdyPersistentException;
 import de.jdynameta.base.value.TypedValueObject;
 import de.jdynameta.base.value.ValueObject;
 import de.jdynameta.persistence.state.ApplicationObj;
 
 import javax.persistence.EntityManager;
+import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.metamodel.EntityType;
+import java.util.ArrayList;
+import java.util.List;
 
 public class JpaWriter {
 
@@ -41,8 +46,35 @@ public class JpaWriter {
 
     public void deleleteInDb(final ApplicationObj objToDelete, EntityType<?> entityType, EntityManager entityManager) throws JdyPersistentException {
 
-        final Object jpaObjToDelete = null;
+        final List<ObjectFilterExpression> idExprList = new ArrayList<>();
 
-        entityManager.remove(jpaObjToDelete);
+        objToDelete.getClassInfo().handleAttributes(new AttributeHandler() {
+            @Override
+            public void handleObjectReference(ObjectReferenceAttributeInfo aInfo, ValueObject objToHandle) throws JdyPersistentException {
+
+                if (aInfo.isKey()) {
+                    idExprList.add(new DefaultObjectReferenceEqualExpression(aInfo,objToHandle));
+                }
+            }
+
+            @Override
+            public void handlePrimitiveAttribute(PrimitiveAttributeInfo aInfo, Object objToHandle) throws JdyPersistentException {
+                if (aInfo.isKey()) {
+                    final DefaultOperatorExpression opExpr = new DefaultOperatorExpression();
+                    opExpr.setAttributeInfo(aInfo);
+                    opExpr.setCompareValue(objToHandle);
+                    opExpr.setMyOperator(DefaultOperatorEqual.getEqualInstance());
+                    idExprList.add(opExpr);
+                }
+
+            }
+        }, objToDelete);
+
+        final DefaultClassInfoQuery query = new DefaultClassInfoQuery(objToDelete.getClassInfo(), new DefaultExpressionAnd(idExprList));
+        final CriteriaQuery<Object> criteriaQuery = new JpaFilterConverter(entityManager).convert( query);
+        final Object jpaObjToDelete = entityManager.createQuery(criteriaQuery).getSingleResult();
+        if (jpaObjToDelete != null) {
+            entityManager.remove(jpaObjToDelete);
+        }
     }
 }
